Changing to agpypeline library support
