+++ generate_docker.py
@@ -4,13 +4,14 @@
 """
 
 import argparse
-import configuration
+from configuration import ConfigurationCanopycover
 
 # The template file name for Dockerfile
 DOCKERFILE_TEMPLATE_FILE_NAMES = ["Dockerfile.template"]
 
 # The default docker image to use
 DEFAULT_DOCKER_IMAGE = 'agpipeline/gantry-base-image:latest'
+
 
 def determine_base_image() -> str:
     """Determines the base image to use in the dockerfile
@@ -26,18 +27,19 @@ def determine_base_image() -> str:
 
     return args.base_image
 
+
 def generate_dockerfile(base_image_name: str) -> None:
-    """Genertes a Dockerfile file using the configured information
+    """Generates a Dockerfile file using the configured information
     """
     # pylint: disable=global-statement
     global DOCKERFILE_TEMPLATE_FILE_NAMES
 
     missing = []
-    if not hasattr(configuration, 'TRANSFORMER_NAME') or not configuration.TRANSFORMER_NAME:
+    if not hasattr(ConfigurationCanopycover, 'transformer_name') or not ConfigurationCanopycover.transformer_name:
         missing.append("Transformer name")
-    if not hasattr(configuration, 'AUTHOR_NAME') or not configuration.AUTHOR_NAME:
+    if not hasattr(ConfigurationCanopycover, 'author_name') or not ConfigurationCanopycover.author_name:
         missing.append("Author name")
-    if not hasattr(configuration, 'AUTHOR_EMAIL') or not configuration.AUTHOR_EMAIL:
+    if not hasattr(ConfigurationCanopycover, 'author_email') or not ConfigurationCanopycover.author_email:
         missing.append("Author email")
     if missing:
         raise RuntimeError("One or more configuration fields aren't defined in configuration.py: " \
@@ -50,12 +52,13 @@ def generate_dockerfile(base_image_name: str) -> None:
         with open(dockerfile_name, 'w') as out_file:
             for line in template:
                 if line.startswith('LABEL maintainer='):
-                    out_file.write("LABEL maintainer=\"{0} <{1}>\"\n".format(configuration.AUTHOR_NAME, \
-                                   configuration.AUTHOR_EMAIL))
+                    out_file.write("LABEL maintainer=\"{0} <{1}>\"\n".format(ConfigurationCanopycover.author_name, \
+                                   ConfigurationCanopycover.author_email))
                 elif line.startswith('FROM base-image'):
                     out_file.write("FROM {0}\n".format(base_image_name))
                 else:
                     out_file.write("{0}\n".format(line))
+
 
 # Make the call to generate the file
 if __name__ == "__main__":
